+++ src/GatewaySettings.php
@@ -7,9 +7,17 @@
  * Company: Pronamic
  *
  * @author Remco Tolsma
- * @version 1.2.4
- * @since 1.2.4
+ * @version 1.3.5
+ * @since 1.3.0
  */
 abstract class Pronamic_WP_Pay_GatewaySettings {
-
+	/**
+	 * Return data for Pronamic_WP_Pay_Admin_GatewayPostType::save_post().
+	 *
+	 * @param array $data
+	 * @return array
+	 */
+	public function save_post( $data ) {
+		return $data;
+	}
 }