+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a CHANGELOG](http://keepachangelog.com/).
 
 ## [Unreleased][unreleased]
--
+
+## [1.3.5] - 2016-03-22
+- Add Pronamic_WP_Pay_GatewaySettings::save_post() to modify data when a gateway is saved.
 
 ## [1.3.4] - 2016-03-02
 - Use the new get_gateway_class() function which is new on the config objects.
@@ -49,7 +51,9 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## 1.0.0
 - First release.
 
-[unreleased]: https://github.com/wp-pay/core/compare/1.3.1...HEAD
+[unreleased]: https://github.com/wp-pay/core/compare/1.3.5...HEAD
+[1.3.5]: https://github.com/wp-pay/core/compare/1.3.4...1.3.5
+[1.3.4]: https://github.com/wp-pay/core/compare/1.3.3...1.3.4
 [1.3.3]: https://github.com/wp-pay/core/compare/1.3.2...1.3.3
 [1.3.2]: https://github.com/wp-pay/core/compare/1.3.1...1.3.2
 [1.3.1]: https://github.com/wp-pay/core/compare/1.3.0...1.3.1